requirements.txt
aiohappyeyeballs==2.6.1
aiohttp==3.13.2
aiosignal==1.4.0
asgiref==3.11.0
asttokens==3.0.1
attrs==25.4.0
certifi==2025.11.12
charset-normalizer==3.4.4
colorama==0.4.6
comm==0.2.3
debugpy==1.8.19
decorator==5.2.1
Django==6.0.1
django-cors-headers==4.9.0
djangorestframework==3.16.1
dotenv==0.9.9
executing==2.2.1
faiss-cpu==1.13.2
filelock==3.20.0
frozenlist==1.8.0
fsspec==2025.12.0
idna==3.11
ipykernel==7.1.0
ipython==9.8.0
ipython_pygments_lexers==1.1.1
jedi==0.19.2
Jinja2==3.1.6
jupyter_client==8.7.0
jupyter_core==5.9.1
MarkupSafe==2.1.5
matplotlib-inline==0.2.1
mpmath==1.3.0
multidict==6.7.0
neo4j==6.0.3
nest-asyncio==1.6.0
networkx==3.6.1
numpy==2.3.5
packaging==25.0
parso==0.8.5
pillow==12.0.0
platformdirs==4.5.1
prompt_toolkit==3.0.52
propcache==0.4.1
psutil==7.2.1
psycopg2-binary==2.9.11
pure_eval==0.2.3
Pygments==2.19.2
pyparsing==3.3.1
python-dateutil==2.9.0.post0
python-dotenv==1.2.1
pytz==2025.2
pyzmq==27.1.0
requests==2.32.5
setuptools==70.2.0
six==1.17.0
sqlparse==0.5.5
stack-data==0.6.3
sympy==1.14.0
torch==2.9.1+cu126
torch-geometric==2.7.0
torchvision==0.24.1+cu126
tornado==6.5.4
tqdm==4.67.1
traitlets==5.14.3
typing_extensions==4.15.0
tzdata==2025.3
urllib3==2.6.2
wcwidth==0.2.14
wheel==0.45.1
xxhash==3.6.0
yarl==1.22.0